Cherrypick deployment changes
